Optimized player movement and changed two animations
The animations were changed just so the Game Over text doesn't appear in the corner of the screen during the gameplay on mobile devices.
Even though the game isn't supposed to be played on mobile, I decided to fix it because it looked really bad and it was easily fixable.